requirements.txt
Flask==3.0.0
Flask-Cors==5.0.0
numpy==1.26.1
pandas==2.1.1
scikit-learn==1.4.2
requests==2.31.0
matplotlib==3.8.4
pymongo==4.6.1
python-dotenv==1.0.1